+++ ui/__init__.py
@@ -4,6 +4,7 @@ from .problems_panel import ProblemsPanel
 from .plugins_dialog import PluginsDialog
 from .shortcuts_dialog import ShortcutsDialog
 from .search_dialog import FindReplaceDialog
+from .diff_viewer import DiffViewerDialog
 
 __all__ = [
     "MainWindow",
@@ -11,4 +12,5 @@ __all__ = [
     "PluginsDialog",
     "ShortcutsDialog",
     "FindReplaceDialog",
+    "DiffViewerDialog",
 ]